Fix gcc issues
Former-commit-id: 47843455d1c6376c25259e48b970b89725f74663

+++ src/gamelib/nlhe_state.cpp
@@ -249,7 +249,7 @@ const nlhe_state* nlhe_state::get_child(int index) const
     if (index == -1)
         return nullptr;
 
-    if (index >= children_.size())
+    if (index >= static_cast<int>(children_.size()))
         return nullptr;
 
     return children_[index].get();
